refactor(conversation): extract approval resolution runtime
